Make a problem set app suited for hosting RTutor in the web

The user first opens the login app, which creates a session file and then calls this app.

RTutorPSApp(ps.name, user.name = "default_user", sample.solution = FALSE,
  run.solved = load.sav, import.rmd = FALSE, rmd.file = paste0(ps.name,
  "_", user.name, "_export.rmd"), catch.errors = TRUE, dir = getwd(),
  rps.dir = dir, ups.dir = paste0(dir, "/ups"),
  offline = !can.connect.to.MathJax(), left.margin = 2, right.margin = 2,
  save.nothing = FALSE, show.solution.btn = TRUE, show.data.exp = FALSE,
  disable.graphics.dev = TRUE, clear.user = FALSE,
  check.whitelist = !is.null(wl), wl = NULL, verbose = FALSE,
  html.data.frame = TRUE, table.max.rows = 25, round.digits = 8,
  signif.digits = 8, knit.print.opts = make.knit.print.opts(html.data.frame
  = html.data.frame, table.max.rows = table.max.rows, round.digits =
  round.digits, signif.digits = signif.digits), precomp = FALSE,
  noeval = FALSE, need.login = TRUE, sessions.dir = paste0(dir,
  "/sessions"), session.key = NULL, use.secure.eval = TRUE,
  secure.eval.timeout = 10, secure.eval.profile = NULL,
  hint.noeval = noeval, show.points = TRUE, replace.sol = precomp,
  ups.save = default.ups.save(chunk.ind = TRUE, code = !(replace.sol |
  noeval)), log.file = paste0(dir, "/log/", ps.name, ".log"),
  session.timeout.sec = 300, ...)

Arguments

sample.solution
shall the sample solution be shown

run.solved
if sample.solution or load.sav shall the correct chunks be automatically run when the problem set is loaded? (Starting the problem set then may take quite a while)

import.rmd
shall the solution be imported from the rmd file specificed in the argument rmd.file

catch.errors
by default TRUE only set FALSE for debugging purposes in order to get a more informative traceback()

offline
(FALSE or TRUE) Do you have no internet connection. By default it is checked whether RTutor can connect to the MathJax server. If you have no internet connection, you cannot render mathematic formulas. If RTutor wrongly thinks you have an internet connection, while you don't, your chunks may not show at all. If you encounter this problem, set manually offline=TRUE.

html.data.frame
shall data.frames and matrices be printed as html table if a chunk is checked? (Default=TRUE)

table.max.rows
the maximum number of rows that is shown if a data.frame is printed as html.table

round.digits
the number of digits that printed data.frames shall be rounded to

load.sav
shall the last saved be loaded?

lauch.browser
if TRUE (default) show the problem set in the browser. Otherwise it is shown in the RStudio viewer pane

is.solved
DEPRECEATED