Run a problem set from a package in the browser

Only works if a package for the problem set is loaded. For problem sets stored in a local .rps file, use show.ps() instead

run.ps(user.name, ps.name = info$ps[1], dir = getwd(), package = NULL,
  auto.save.code = FALSE, clear.user = FALSE, run.solved = FALSE,
  sample.solution = FALSE, show.solution.btn = NA, launch.browser = TRUE,
  info = get.package.info(package), deploy.local = !make.web.app,
  make.web.app = FALSE, pkg.dir = path.package(info$package),
  rps.dir = find.pkg.rps.dir(ps.name, pkg.dir),
  material.dir = find.pkg.material.dir(ps.name, pkg.dir), ...)

Arguments

user.name
Your user name

ps.name
Name of the problem set. By default the first problem set name of the loaded RTutor problem set package.

dir
your working directory for the problem set

package
name of the package that contains your problem set. Is automatically chosen a (single) package with an RTutor problem set is loaded.

auto.save.code
If TRUE all entered code will be automatically saved for the user. If FALSE (default) no entered code is saved. The user statistics (how many chunks are solved) are still saved, however.

clear.user
If TRUE all previously saved data for the user is removed if the problem set starts. Can be useful for developlmen or for resetting things.

run.solved
If TRUE and also sample.soulution=TRUE all previously solved chunks will be run when the problem set is newly shown. This can be very time consuming. I would suggest in most cases to keep the default run.solved=FALSE.

sample.solution
If TRUE the sample solution is shown in all chunks. Can be useful when developing a problem set. Note that one can create a problem set such that the sample solution is not available, e.g. if one wants to avoid that students just look at the sample solution.

show.solution.btn
If TRUE add a button to each chunk to show the sample solution. Note that one can create a problem set such that the sample solution is never available.

launch.browser
if TRUE (default) show the problem set in the browser. Otherwise it is shown in the RStudio viewer pane

pkg.dir
the package directory under which problem set files are searched under pkg.dir/ps/ps.name/. Will be set by default to currently loaded RTutorProblemSet package

rps.dir
directory of rps.files. Will be set to default for current package

material.dir
directory of additional problem set files. Will be set to default for current package

...
additional arguments of show.ps